requirements.txt
asgiref==3.4.1
bootstrap4==0.1.0
boto3==1.20.37
botocore==1.23.37
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
colorama==0.4.4
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.1
django-allauth==0.47.0
django-allauth-2fa==0.8
django-bootstrap3==21.2
django-otp==1.1.3
django-storages==1.12.3
gcloud==0.18.3
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.3
jmespath==0.10.0
jwcrypto==1.0
messagebird==2.0.0
mysqlclient==2.1.0
oauth2client==4.1.3
oauthlib==3.1.1
Pillow==9.0.0
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.12.0
PyJWT==2.3.0
pyparsing==3.0.6
Pyrebase4==4.5.0
python-dateutil==2.8.2
python-jwt==3.3.2
python3-openid==3.2.0
pytz==2021.3
qrcode==7.3.1
requests==2.27.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.8
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
twilio==7.5.0
tzdata==2021.5
urllib3==1.26.8
wrapt==1.13.3
